Converted project to maven
