+++ Plugin/src/main/java/de/Linus122/TimeIsMoney/Metrics.java
