+++ Tools/src/main/java/de/Linus122/TimeIsMoney/tools/ActionBarUtils.java
@@ -1,0 +1,7 @@
+package de.Linus122.TimeIsMoney.tools;
+
+import org.bukkit.entity.Player;
+
+public interface ActionBarUtils {
+	void sendActionBarMessage(Player p, String message);
+}
